requirements.txt
python-telegram-bot==13.15
requests==2.31.0
python-dotenv==1.0.1
urllib3<2
six==1.16.0
